Change approach to instrumentation of statements
Previously, we traversed the whole AST, and when we encountered a
statement we checked if we could instrument it i.e. if it was
standalone. This introduced a bunch of cases that we wanted to check
(is it in a block? it is immediately after a block? etc.) and some edge
cases that didn’t work properly (comments).

Instead, this stops traversing the AST in situations where we know we
can’t instrument. We can’t instrument `f(x)` easily (see below) in
these situations, so we don’t traverse them:

* `if (f(x))`
* `arr[f(x)`
*  `var x = f(x)`
* `g(f(x))`

Of course, going forward we might ‘decide’ that we *can* in fact
instrument these. e.g. `arr[f(x)]` could be replaced by

```
var _some_unique_id = f(x);
arr[_some_unique_id];
```
and then instrumented as appropriate.